refactor: use `findOverlapCodeRange` and move it to `@volar/language-core`
