+++ packages/language-service/lib/features/provideDocumentFormattingEdits.ts
@@ -1,10 +1,10 @@
-import { SourceScript, VirtualCode, forEachEmbeddedCode, isFormattingEnabled } from '@volar/language-core';
+import { SourceScript, VirtualCode, findOverlapCodeRange, forEachEmbeddedCode, isFormattingEnabled } from '@volar/language-core';
 import type * as vscode from 'vscode-languageserver-protocol';
 import { TextDocument } from 'vscode-languageserver-textdocument';
 import { URI } from 'vscode-uri';
 import type { EmbeddedCodeFormattingOptions, LanguageServiceContext } from '../types';
 import { NoneCancellationToken } from '../utils/cancellation';
-import { findOverlapCodeRange, stringToSnapshot } from '../utils/common';
+import { stringToSnapshot } from '../utils/common';
 import { DocumentsAndMap, getGeneratedPositions, getSourceRange } from '../utils/featureWorkers';
 
 export function register(context: LanguageServiceContext) {
