+++ packages/language-service/lib/utils/common.ts
@@ -1,75 +1,5 @@
-import type { CodeInformation, Mapper } from '@volar/language-core';
 import type * as ts from 'typescript';
 import type * as vscode from 'vscode-languageserver-protocol';
-
-export function findOverlapCodeRange(
-	start: number,
-	end: number,
-	map: Mapper,
-	filter: (data: CodeInformation) => boolean
-) {
-	let mappedStart: number | undefined;
-	let mappedEnd: number | undefined;
-
-	for (const [mapped, mapping] of map.toGeneratedLocation(start)) {
-		if (filter(mapping.data)) {
-			mappedStart = mapped;
-			break;
-		}
-	}
-	for (const [mapped, mapping] of map.toGeneratedLocation(end)) {
-		if (filter(mapping.data)) {
-			mappedEnd = mapped;
-			break;
-		}
-	}
-
-	if (mappedStart === undefined || mappedEnd === undefined) {
-		for (const mapping of map.mappings) {
-			if (filter(mapping.data)) {
-				const mappingStart = mapping.sourceOffsets[0];
-				const mappingEnd = mapping.sourceOffsets[mapping.sourceOffsets.length - 1] + mapping.lengths[mapping.lengths.length - 1];
-				const overlap = getOverlapRange(start, end, mappingStart, mappingEnd);
-				if (overlap) {
-					const curMappedStart = (overlap.start - mappingStart) + mapping.generatedOffsets[0];
-					const lastGeneratedLength = (mapping.generatedLengths ?? mapping.lengths)[mapping.generatedOffsets.length - 1];
-					const curMappedEndOffset = Math.min(overlap.end - mapping.sourceOffsets[mapping.sourceOffsets.length - 1], lastGeneratedLength);
-					const curMappedEnd = mapping.generatedOffsets[mapping.generatedOffsets.length - 1] + curMappedEndOffset;
-
-					mappedStart = mappedStart === undefined ? curMappedStart : Math.min(mappedStart, curMappedStart);
-					mappedEnd = mappedEnd === undefined ? curMappedEnd : Math.max(mappedEnd, curMappedEnd);
-				}
-			}
-		}
-	}
-
-	if (mappedStart !== undefined && mappedEnd !== undefined) {
-		return {
-			start: mappedStart,
-			end: mappedEnd,
-		};
-	}
-}
-
-function getOverlapRange(
-	range1Start: number,
-	range1End: number,
-	range2Start: number,
-	range2End: number
-): { start: number, end: number; } | undefined {
-
-	const start = Math.max(range1Start, range2Start);
-	const end = Math.min(range1End, range2End);
-
-	if (start > end) {
-		return undefined;
-	}
-
-	return {
-		start,
-		end,
-	};
-}
 
 export function isInsideRange(parent: vscode.Range, child: vscode.Range) {
 	if (child.start.line < parent.start.line) {
